code_reviewer/requirements.txt
python-dotenv
google-generativeai
langchain
langchain-google-genai
openai
langchain_community